update settings page
