--- web/src/pages/settings/BrandingCard/styles.scss
@@ -1,44 +1,0 @@
-@use '../../../shared/scss/helpers' as *;
-
-#settings-page {
-  .branding {
-    @include media-breakpoint-up(lg) {
-      margin-bottom: 30px;
-    }
-
-    & > .card {
-      box-sizing: border-box;
-      padding: 25px 20px 25px 15px;
-
-      & > form {
-        align-items: stretch;
-        justify-content: flex-start;
-
-        .helper {
-          margin-left: auto;
-        }
-      }
-
-      & > header {
-        display: flex;
-        flex-flow: row nowrap;
-        align-content: center;
-        align-items: flex-start;
-        justify-content: flex-start;
-        width: 100%;
-        margin-bottom: 15px;
-
-        & > h3 {
-          @include small-header;
-        }
-
-        & > .controls {
-          display: flex;
-          flex-flow: row;
-          margin-left: auto;
-          gap: 10px;
-        }
-      }
-    }
-  }
-}
