--- web/src/pages/settings/DefaultNetworkSelect/style.scss
@@ -1,20 +1,0 @@
-@use '../../../shared/scss/helpers' as *;
-
-#settings-page {
-  .network-view {
-    @include media-breakpoint-up(lg) {
-      margin-bottom: 50px;
-    }
-
-    & > header {
-      display: flex;
-      flex-flow: row nowrap;
-      column-gap: 10px;
-    }
-
-    & > .select {
-      height: 40px;
-      width: 140px;
-    }
-  }
-}
