--- web/src/pages/settings/ModulesCard/styles.scss
@@ -1,27 +1,0 @@
-@use '../../../shared/scss/helpers' as *;
-
-#settings-page {
-  .modules {
-    @include media-breakpoint-up(lg) {
-      margin-bottom: 30px;
-    }
-
-    & > .card {
-      display: flex;
-      flex-flow: column;
-      box-sizing: border-box;
-      padding: 12px 15px 20px;
-      row-gap: 15px;
-
-      .checkbox {
-        & > .label {
-          color: var(--gray-dark);
-
-          & > * {
-            color: var(--gray-dark);
-          }
-        }
-      }
-    }
-  }
-}
