+++ web/src/pages/settings/SettingsPage.tsx
@@ -1,73 +1,60 @@
 import './style.scss';
 
-import { useState } from 'react';
-import { useBreakpoint } from 'use-breakpoint';
+import { useQueryClient } from '@tanstack/react-query';
+import { ReactNode, useEffect, useMemo, useState } from 'react';
 
 import { useI18nContext } from '../../i18n/i18n-react';
 import { PageContainer } from '../../shared/components/Layout/PageContainer/PageContainer';
-import { deviceBreakpoints } from '../../shared/constants';
 import { Card } from '../../shared/defguard-ui/components/Layout/Card/Card';
 import { CardTabs } from '../../shared/defguard-ui/components/Layout/CardTabs/CardTabs';
+import { CardTabsData } from '../../shared/defguard-ui/components/Layout/CardTabs/types';
 import { useAppStore } from '../../shared/hooks/store/useAppStore';
-import { BrandingCard } from './BrandingCard/BrandingCard';
-import { BuiltByCard } from './BuiltByCard/BuiltByCard';
-import { ModulesCard } from './ModulesCard/ModulesCard';
-import { SmtpTab } from './SmtpTab/SmtpTab';
-import { SupportCard } from './SupportCard/SupportCard';
-import { Web3Settings } from './Web3Settings/Web3Settings';
+import { QueryKeys } from '../../shared/queries';
+import { GlobalSettings } from './components/GlobalSettings/GlobalSettings';
+import { SmtpSettings } from './components/SmtpSettings/SmtpSettings';
 
-enum Tabs {
-  Basic,
-  Smtp,
-}
+const tabsContent: ReactNode[] = [<GlobalSettings key={0} />, <SmtpSettings key={1} />];
 
 export const SettingsPage = () => {
+  const queryClient = useQueryClient();
   const { LL } = useI18nContext();
-  const [tab, setTab] = useState(Tabs.Basic);
-  const tabs = [
-    {
-      key: 1,
-      onClick: () => {
-        setTab(Tabs.Basic);
+
+  const [activeCard, setActiveCard] = useState(0);
+
+  const tabs = useMemo(
+    (): CardTabsData[] => [
+      {
+        key: 0,
+        content: LL.settingsPage.tabs.global(),
+        active: activeCard === 0,
+        onClick: () => setActiveCard(0),
       },
-      content: LL.settingsPage.tabs.general(),
-      active: tab === Tabs.Basic,
-    },
-    {
-      key: 2,
-      onClick: () => {
-        setTab(Tabs.Smtp);
+      {
+        key: 1,
+        content: LL.settingsPage.tabs.smtp(),
+        active: activeCard === 1,
+        onClick: () => setActiveCard(1),
       },
-      content: LL.settingsPage.tabs.smtp(),
-      active: tab === Tabs.Smtp,
-    },
-  ];
+    ],
+    [LL.settingsPage.tabs, activeCard],
+  );
+
   const settings = useAppStore((state) => state.settings);
-  const { breakpoint } = useBreakpoint(deviceBreakpoints);
+
+  // Refetch settings on page mount
+  useEffect(() => {
+    queryClient.invalidateQueries([QueryKeys.FETCH_SETTINGS]);
+    // eslint-disable-next-line
+  }, []);
+
+  if (!settings) return null;
+
   return (
     <PageContainer id="settings-page">
-      <header>
-        <h1>
-          {settings?.instance_name} {LL.settingsPage.title()}
-        </h1>
-      </header>
-      {breakpoint === 'desktop' && <CardTabs tabs={tabs} />}
-      <Card className="settings-card" hideMobile>
-        {tab === Tabs.Basic && (
-          <>
-            <div className="left">
-              <BrandingCard />
-              <ModulesCard />
-              {/*<DefaultNetworkSelect /> */}
-            </div>
-            <div className="right">
-              <Web3Settings />
-              <SupportCard />
-              <BuiltByCard />
-            </div>
-          </>
-        )}
-        {tab === Tabs.Smtp && <SmtpTab />}
+      <h1>{LL.settingsPage.title()}</h1>
+      <CardTabs tabs={tabs} />
+      <Card className="settings-card" hideMobile shaded>
+        {tabsContent[activeCard]}
       </Card>
     </PageContainer>
   );
