--- web/src/pages/settings/Web3Settings/style.scss
@@ -1,43 +1,0 @@
-@use '../../../shared/scss/helpers' as *;
-
-#settings-page {
-  .web3-settings {
-    & > .card {
-      box-sizing: border-box;
-      padding: 25px 20px 25px 15px;
-
-      & > header {
-        display: flex;
-        flex-flow: row nowrap;
-        align-content: center;
-        align-items: flex-start;
-        justify-content: flex-start;
-        width: 100%;
-        margin-bottom: 15px;
-
-        & > h3 {
-          @include small-header;
-        }
-
-        & > .controls {
-          margin-left: auto;
-        }
-      }
-
-      & > textarea {
-        resize: none;
-        width: 100%;
-        max-width: 100%;
-        overflow-x: hidden;
-        height: 147px;
-        margin: 0;
-        box-sizing: border-box;
-        padding: 20px 15px;
-        border-radius: 10px;
-        border: 1px solid var(--gray-lighter);
-
-        @include typography-legacy(15px, 23px, regular, var(--text-main), 'Poppins');
-      }
-    }
-  }
-}
