+++ web/src/pages/settings/components/GlobalSettings/GlobalSettings.tsx
@@ -1,0 +1,15 @@
+import { BrandingSettings } from './components/BrandingSettings/BrandingSettings';
+import { ModulesSettings } from './components/ModulesSettings/ModulesSettings';
+import { Web3Settings } from './components/Web3Settings/Web3Settings';
+
+export const GlobalSettings = () => (
+  <>
+    <div className="left">
+      <BrandingSettings />
+      <ModulesSettings />
+    </div>
+    <div className="right">
+      <Web3Settings />
+    </div>
+  </>
+);
