+++ web/src/pages/settings/components/GlobalSettings/components/BrandingSettings/styles.scss
@@ -1,0 +1,20 @@
+@use '@scssutils' as *;
+
+#branding-settings {
+  & > .card {
+    @include media-breakpoint-up(lg) {
+      padding: 17px 15px 5px 17px;
+    }
+
+    & > .controls {
+      margin-bottom: 20px;
+    }
+
+    & > form {
+      width: 100%;
+      & > * {
+        width: 100%;
+      }
+    }
+  }
+}
