+++ web/src/pages/settings/components/SmtpSettings/components/SmtpEncryption/style.scss
@@ -1,0 +1,7 @@
+@use '@scssutils' as *;
+
+#smtp-encryption {
+  .select-container {
+    width: 150px;
+  }
+}
