+++ web/src/pages/settings/components/SmtpSettings/styles.scss
@@ -1,0 +1,1 @@
+@use '@scssutils' as *;
