+++ web/src/pages/support/SupportPage.tsx
@@ -1,0 +1,3 @@
+export const SupportPage = () => {
+  return <></>;
+};
