+++ web/src/pages/support/components/BuiltByCard/BuiltByCard.tsx
