+++ web/src/pages/support/components/BuiltByCard/style.scss
