+++ web/src/pages/support/components/SupportCard/style.scss
